Version with scale-down and add-up option
In this version, if the target ice concentration is less than the current amount,
the sea ice distribution is rescaled to be closer to the target.
If the the target ice concentration is greater than the current amount, then
the concentration is increased in each category up to the thickest category that
currently has ice. If ice is added to a category ice that had no ice, the thickness
is set to the minimum thickness bound of that category. The salinity and enthalypy
of the newly created ice is also adjusted.
This version still includes a number of debugging statements.