Add check to skip nudging if the target and the current concentraion are similar
When the target and the current ice concentrations were too similar, I was getting
errors. I think from subtracting two numbers very close to 1 and dividing.
I fixed this by adding a check before the nudging is applied to do nothing
if the target and current concentrations are the same(ish).